--- a/dat/constants.lua
+++ b/dat/constants.lua
@@ -1,6 +1,6 @@
 return {
    -- Constants used by the Engine
-   PHYSICS_SPEED_DAMP = 1 / 0.15, -- Was 3 until 0.13.0
+   PHYSICS_SPEED_DAMP = 4.5, -- Was 3 until 0.13.0
 
    -- Constants used by Lua scripts
    BITE_ACCEL_MOD = 500,