--- a/dat/outfits/bioship/generate.py
+++ b/dat/outfits/bioship/generate.py
@@ -145,7 +145,7 @@ typename["hull"] = N_("Bioship Shell")
 # Mediocris Gene Drive  =>  Tricon Cyclone
 # Largus Gene Drive  =>  Tricon Cyclone II
 # Ponderosus Gene Drive  =>  Tricon Typhoon
-# Immanis Gene Drive  =>  Eagle 6500
+# Immanis Gene Drive  =>  Bolt 6500
 # Magnus Gene Drive  =>  Tricon Typhoon2
 # Grandis Gene Drive  =>  Melendez Mammoth
 for nam,temp,gfx,output_pref,outputs in [
@@ -155,7 +155,7 @@ for nam,temp,gfx,output_pref,outputs in [
    ("medium/tricon_cyclone_engine.xml","gene_drive_tricon","fast_m","Mediocris",["I","II","III"]),
    ("medium/tricon_cyclone_ii_engine.xml","gene_drive_tricon","strong_m","Largus",["I","II"]),
    ("large/tricon_typhoon_engine.xml","gene_drive_tricon","fast_l","Ponderosus",["I","II","III"]),
-   ("large/unicorp_eagle_6500_engine.xml","gene_drive","strong_l","Immanis",["I","II","III"]),
+   ("large/nexus_bolt_6500_engine.xml","gene_drive","strong_l","Immanis",["I","II","III"]),
    ("large/tricon_typhoon_ii_engine.xml","gene_drive_tricon","strong_l","Magnus",["I","II","III"]),
    ("large/melendez_old_mammoth_engine.xml","gene_drive_melendez","strong_l","Grandis",["I","II","III"]),
 ]: