--- a/dat/scripts/bioship/skills.lua
+++ b/dat/scripts/bioship/skills.lua
@@ -71,7 +71,7 @@ skills.set.move = {
       test = function( p )
          for k,o in ipairs(p:outfitsList()) do
             if o:typeBroad()=="Afterburner" then
-               --if luatk.yesno( _("Remove Afterburner?"), _("Your ship already has an afterburner. Remove to learn the new skill?")A)
+               --if luatk.yesno( _("Remove Afterburner?"), _("Your ship already has an afterburner. Remove to learn the new skill?"))
                luatk.msg(_("Afterburner"),_("Your ship already has an afterburner equipped! Please remove to be able to learn this skill."))
                return false
             end