--- a/dat/scripts/equipopt/optimize.lua
+++ b/dat/scripts/equipopt/optimize.lua
@@ -125,7 +125,7 @@ end
 special_ships["Heavy Drone"] = function( p )
    for k,o in ipairs{
       "Milspec Thalos 3602 Core System",
-      "Unicorp Hawk 360 Engine",
+      "Nexus Dart 360 Engine",
       choose_one{"Nexus Light Stealth Plating", "S&K Light Combat Plating"},
       "Shatterer Launcher",
       "Shatterer Launcher",