--- a/dat/scripts/equipopt/templates/pirate.lua
+++ b/dat/scripts/equipopt/templates/pirate.lua
@@ -64,7 +64,7 @@ local pirate_cores = {
       end,
    ["Pirate Starbridge"] = function (p)
          local c = ecores.get( p, { systems=pirate_class, hulls=pirate_class } )
-         table.insert( c, choose_one{ "Unicorp Falcon 1400 Engine", "Krain Patagium Engine", } )
+         table.insert( c, choose_one{ "Nexus Arrow 1400 Engine", "Krain Patagium Engine", } )
          return c
       end,
 }