--- a/utils/outfits/apply_engines.py
+++ b/utils/outfits/apply_engines.py
@@ -10,6 +10,7 @@ from getconst import PHYSICS_SPEED_DAMP
 AG_EXP  = 0.4
 TURN_CT = 0.40
 STD_R   = 0.15
+R_MAG   = 1.8
 
 sizes={
    "Za'lek Test Engine":2,
@@ -77,8 +78,8 @@ line_stats = {
         "speed" : 1.0, # 1.0 indicates current speed rank, lower means slower, higher means faster
     },
     "K" : {
-        "ratio" : 0.9, # lower ratio
-        "speed" : 1.2, # higher speed than average
+        "ratio" : 1.0, # lower ratio
+        "speed" : 1.3, # higher speed than average
     },
     "N" : {
         "ratio" : 1.0,
@@ -90,19 +91,19 @@ line_stats = {
     },
     "U" : {
         "ratio" : 1.0,
-        "speed" : 0.8,
+        "speed" : 0.7,
     },
     "Z" : { # TODO make these change over time the profile via Lua
         "ratio" : 1.1,
-        "speed" : 0.7,
+        "speed" : 0.65,
     },
     "B" : {
         "ratio" : 1.0,
-        "speed" : 0.7,
+        "speed" : 0.6,
     },
 }
 
-ALPHA, BETA = 1.11, 0.06
+ALPHA, BETA = 1.14, 0.048
 
 def dec_i(n):
    if n<=1:
@@ -126,7 +127,7 @@ def ls2vals(line_size):
    fullspeed = dec( size + 1.0 - stats["speed"])
 
    # r ranges from 15% / 2 (size 6) to 15% * 2 (size 1)
-   r = STD_R * pow(2,-((size-1)-2.5)/2.5)
+   r = STD_R * pow(2,-R_MAG*((size-1)-2.5)/5)
 
    # Modulate ratio based on outfit
    r *= line_stats[line]["ratio"]