--- a/utils/outfits/engine_stats.py
+++ b/utils/outfits/engine_stats.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 from getconst import PHYSICS_SPEED_DAMP