… dat/naevpedia/outfits/meson.build
… dat/outfits/bioship/skills/adrenal_gland_i.xml
… dat/outfits/bioship/skills/adrenal_gland_ii.xml
… dat/outfits/bioship/skills/adrenal_gland_iii.xml
… dat/outfits/core_engine/large/beat_up_large_engine.xml
… dat/outfits/core_engine/large/krain_remige_engine.xml
… dat/outfits/core_engine/large/melendez_mammoth_engine.xml
… dat/outfits/core_engine/large/melendez_mammoth_xl_engine.xml
… dat/outfits/core_engine/large/melendez_old_mammoth_engine.xml
… dat/outfits/core_engine/large/nexus_bolt_3000_engine.xml
… dat/outfits/core_engine/large/nexus_bolt_6500_engine.xml
… dat/outfits/core_engine/large/tricon_typhoon_engine.xml
… dat/outfits/core_engine/large/tricon_typhoon_ii_engine.xml
… dat/outfits/core_engine/large/unicorp_eagle_3000_engine.xml
… dat/outfits/core_engine/large/unicorp_eagle_6500_engine.xml
… dat/outfits/core_engine/medium/beat_up_medium_engine.xml
… dat/outfits/core_engine/medium/krain_patagium_engine.xml
… dat/outfits/core_engine/medium/melendez_buffalo_engine.xml
… dat/outfits/core_engine/medium/melendez_buffalo_xl_engine.xml
… dat/outfits/core_engine/medium/nexus_arrow_1400_engine.xml
… dat/outfits/core_engine/medium/nexus_arrow_700_engine.xml
… dat/outfits/core_engine/medium/tricon_cyclone_engine.xml
… dat/outfits/core_engine/medium/tricon_cyclone_ii_engine.xml
… dat/outfits/core_engine/medium/unicorp_falcon_1400_engine.xml
… dat/outfits/core_engine/medium/unicorp_falcon_700_engine.xml
… dat/outfits/core_engine/small/beat_up_small_engine.xml
… dat/outfits/core_engine/small/melendez_ox_engine.xml
… dat/outfits/core_engine/small/melendez_ox_xl_engine.xml
… dat/outfits/core_engine/small/nexus_dart_160_engine.xml
… dat/outfits/core_engine/small/nexus_dart_360_engine.xml