**Balance**

This PR addresses the PR #2786.

We focus on balancing the engines mobility parameters (speed,accel,turn) and the physical constant PHYSICS\_SPEED\_DAMP in order to have a satisfying piloting look and feel. In particular, the different engine lines and size should feel different.

## Changes

### Speed scale

It has been reduced a bit : it now ranges from 100 to 400 for the Tricon line.
It becomes possible to become as fast as the ship of lesser size, which were previously not the case for larger ships.

### Agility variations
The most obvious change is that the part of maximum speed that is caused by accel is not constant anymore. This part is the difference between 100% and what you get when you stop thrusting. We call this _ratio_ and use it as a measure of _agility_. As I made turning depends on agility, ships with higher agility can take tighter curves. More agility also means (almost by definition) less time to reach full speed.

This ratio was constant so far (15%) and **now varies with ship size and engine line**.
There is an **increase for small ships and Tricon** engines, and an a **decrease for larger ships and Melendez** engines.
<details><summary>Details</summary>The base ratio varies roughly from 7.5% for size 6 ships to 30% for size 1 ships.
The base ratio gets multiplied by a factor depending on the ship line : around 1.5 for Tricon, 0.7 for Melendez and 1.0 for Nexus and Unicorp.</details>

### Lines changes
Here I **emphasize** changes.

The **engine lines** have been completed for all sizes (as a requirement for future engine merging) except Krain (that would make no sense to extend).

**Nexus and Unicorp** are now **different**. Unicorp is the default, base line.

The following lines are high quality engines that share similar top speed:
 - **Tricon**: mainly characterized by its **high agility** and low engine mass limit, also slightly faster than the others
 - **Melendez**: mainly characterized by its low agility (therefore its better drift speed) and high engine mass limit, it **is no longer a line of slow engines**.
 - Nexus: All-round line with no downside almost as fast as the others.

Unicorp is an all-round average max speed line.

Krain is a rather balanced line that has top speed intermediate between its size and the size before, making it the fastest for their size. (which is not new).

### Acceleration of small ships
Small ships, and in particular small ships with Tricon, see their acceleration rates grow significantly. First, it gives a feeling of responsiveness that is pleasant to experience.

Second, it makes acceleration a significant parameter, which it was not. Consequently, all outfits/ships acceleration modifiers that were huge (afterburners, Hyena bonuses, etc.) were significantly reduced. Afterburners now rely a bit more on speed and a bit less on acceleration, which allow them to be also useful with medium ships.

### Before / After
<details><summary>Before</summary>

| | speed | max speed | accel | fullsp (s) | fullsp (km) | turn | turn radius | 1/2 turn (s)
| ---                   | ---    | ---    | ---    | ---    | ---    | ---    | ---    | ---
| T. Zephyr Engine      | 345    | 410    | 195    |   2.1  | 431    | 160    | 147    |   1.12
| N. Dart 160 Engine    | 300    | 355    | 165    |   2.15 | 382    | 140    | 145    |   1.29
| T. Zephyr II Engine   | 290    | 343.33 | 160    |   2.15 | 368    | 140    | 141    |   1.29
| U. Hawk 360 Engine    | 250    | 296.67 | 140    |   2.12 | 314    | 120    | 142    |   1.5
| T. Cyclone Engine     | 230    | 273.33 | 130    |   2.1  | 287    | 115    | 136    |   1.57
| M. Ox XL Engine       | 225    | 263.33 | 115    |   2.29 | 301    |  95    | 159    |   1.89
| N. Arrow 700 Engine   | 200    | 238.33 | 115    |   2.07 | 247    | 100    | 137    |   1.8
| K. Patagium Engine    | 190    | 228.33 | 115    |   1.99 | 227    | 100    | 131    |   1.8
| T. Cyclone II Engine  | 175    | 208.33 | 100    |   2.08 | 217    |  90    | 133    |   2
| Z. Test Engine        | 175    | 208.33 | 100    |   2.08 | 217    |  85    | 140    |   2.12
| U. Falcon 1400 Engine | 150    | 176.67 |  80    |   2.21 | 195    |  80    | 127    |   2.25
| Beat Up Small Engine  | 150    | 176.67 |  80    |   2.21 | 195    |  65    | 156    |   2.77
| M. Buffalo XL Engine  | 135    | 158.33 |  70    |   2.26 | 179    |  65    | 140    |   2.77
| K. Remige Engine      | 125    | 148.33 |  70    |   2.12 | 157    |  80    | 106    |   2.25
| T. Typhoon Engine     | 115    | 136.67 |  65    |   2.1  | 144    |  75    | 104    |   2.4
| Beat Up Medium Engine | 105    | 125    |  60    |   2.08 | 130    |  55    | 130    |   3.27
| N. Bolt 3000 Engine   | 100    | 119    |  57    |   2.09 | 124    |  65    | 105    |   2.77
| T. Typhoon II Engine  |  80    |  96.67 |  50    |   1.93 |  93    |  60    |  92    |   3
| U. Eagle 6500 Engine  |  70    |  82.33 |  37    |   2.23 |  92    |  45    | 105    |   4
| M. Mammoth XL Engine  |  65    |  76.67 |  35    |   2.19 |  84    |  45    |  98    |   4
| Beat Up Large Engine  |  55    |  65    |  30    |   2.17 |  70    |  40    |  93    |   4.5
</details><details><summary>After</summary>

| | speed | max speed | accel | fullsp (s) | fullsp (km) | turn | turn radius | 1/2 turn (s)
| ---                       | ---    | ---    | ---    | ---    | ---    | ---    | ---    | ---
| T. Zephyr Engine          | 287    | 400.11 | 509    |   0.79 | 157    | 213    | 108    |   0.85
| M. Ox Engine              | 343    | 394.78 | 233    |   1.69 | 334    | 174    | 130    |   1.03
| N. Dart 160 Engine        | 313    | 392.11 | 356    |   1.1  | 216    | 192    | 117    |   0.94
| Unicorp Hawk 160 Engine   | 298    | 373.33 | 339    |   1.1  | 206    | 183    | 117    |   0.98
| Beat Up Small Engine      | 304    | 354    | 225    |   1.57 | 278    | 159    | 128    |   1.13
| T. Zephyr II Engine       | 259    | 336.33 | 348    |   0.97 | 163    | 170    | 113    |   1.06
| M. Ox XL Engine           | 296    | 331.33 | 159    |   2.08 | 345    | 138    | 138    |   1.3
| N. Dart 360 Engine        | 275    | 329    | 243    |   1.35 | 223    | 153    | 123    |   1.18
| U. Hawk 360 Engine        | 259    | 309.67 | 228    |   1.36 | 210    | 144    | 123    |   1.25
| Z. Test Engine            | 248    | 302.67 | 246    |   1.23 | 186    | 144    | 120    |   1.25
| T. Cyclone Engine         | 222    | 272.89 | 229    |   1.19 | 163    | 131    | 119    |   1.37
| M. Buffalo Engine         | 245    | 268.11 | 104    |   2.58 | 346    | 106    | 145    |   1.7
| N. Arrow 700 Engine       | 230    | 265.33 | 159    |   1.67 | 221    | 117    | 130    |   1.54
| U. Falcon 700 Engine      | 214    | 246.89 | 148    |   1.67 | 206    | 109    | 130    |   1.65
| Beat Up Medium Engine     | 207    | 228.33 |  96    |   2.38 | 272    |  92    | 142    |   1.96
| K. Patagium Engine        | 201    | 228.11 | 122    |   1.87 | 213    | 108    | 121    |   1.67
| T. Cyclone II Engine      | 180    | 212.22 | 145    |   1.46 | 155    |  97    | 125    |   1.86
| M. Buffalo XL Engine      | 193    | 207.67 |  66    |   3.15 | 327    |  78    | 153    |   2.31
| N. Arrow 1400 Engine      | 183    | 205.22 | 100    |   2.05 | 211    |  86    | 137    |   2.09
| U. Falcon 1400 Engine     | 169    | 189.44 |  92    |   2.06 | 195    |  79    | 137    |   2.28
| K. Remige Engine          | 157    | 173.67 |  75    |   2.32 | 201    |  78    | 128    |   2.31
| T. Typhoon Engine         | 140    | 159.56 |  88    |   1.81 | 145    |  69    | 132    |   2.61
| M. Mammoth Engine         | 147    | 155.89 |  40    |   3.9  | 304    |  56    | 159    |   3.21
| M. Old Mammoth Engine     | 147    | 155.89 |  40    |   3.9  | 304    |  56    | 159    |   3.21
| N. Bolt 3000 Engine       | 140    | 153.56 |  61    |   2.52 | 193    |  61    | 144    |   2.95
| Unicorp Eagle 3000 Engine | 128    | 140.22 |  55    |   2.55 | 179    |  56    | 143    |   3.21
| Beat Up Large Engine      | 119    | 126.78 |  35    |   3.62 | 230    |  46    | 158    |   3.91
| T. Typhoon II Engine      | 104    | 115.56 |  52    |   2.22 | 128    |  47    | 141    |   3.83
| M. Mammoth XL Engine      | 107    | 112.11 |  23    |   4.87 | 273    |  38    | 169    |   4.74
| N. Bolt 6500 Engine       | 103    | 111    |  36    |   3.08 | 171    |  42    | 151    |   4.29
| U. Eagle 6500 Engine      |  93    | 100.11 |  32    |   3.13 | 157    |  38    | 151    |   4.74
</details>

### Individual diff
Some files are marked with "\_" to mean "no change" because the file is new.
<details><summary>Engines individual changes</summary>

```
unicorp_falcon_1400_engine.xml: speed:150->169, accel:80->92, turn:80->79
tricon_cyclone_ii_engine.xml: speed:175->180, accel:100->145, turn:90->97
nexus_arrow_700_engine.xml: speed:200->230, accel:115->159, turn:100->117
nexus_arrow_1400_engine.xml: _
melendez_buffalo_xl_engine.xml: speed:135->193, accel:70->66, turn:65->78
beat_up_medium_engine.xml: speed:105->207, accel:60->96, turn:55->92
krain_patagium_engine.xml: speed:190->201, accel:115->122, turn:100->108
unicorp_falcon_700_engine.xml: _
tricon_cyclone_engine.xml: speed:230->222, accel:130->229, turn:115->131
melendez_buffalo_engine.xml: _
krain_remige_engine.xml: speed:125->157, accel:70->75, turn:80->78
melendez_mammoth_engine.xml: _
melendez_old_mammoth_engine.xml: _
melendez_mammoth_xl_engine.xml: speed:65->107, accel:35->23, turn:45->38
unicorp_eagle_6500_engine.xml: speed:70->93, accel:37->32, turn:45->38
beat_up_large_engine.xml: speed:55->119, accel:30->35, turn:40->46
unicorp_eagle_3000_engine.xml: _
tricon_typhoon_engine.xml: speed:115->140, accel:65->88, turn:75->69
nexus_bolt_6500_engine.xml: _
tricon_typhoon_ii_engine.xml: speed:80->104, accel:50->52, turn:60->47
nexus_bolt_3000_engine.xml: speed:100->140, accel:57->61, turn:65->61
unicorp_hawk_360_engine.xml: speed:250->259, accel:140->228, turn:120->144
nexus_dart_360_engine.xml: _
beat_up_small_engine.xml: speed:150->304, accel:80->225, turn:65->159
melendez_ox_xl_engine.xml: speed:225->296, accel:115->159, turn:95->138
unicorp_hawk_160_engine.xml: _
melendez_ox_engine.xml: _
zalek_test_engine.xml: accel:100->246, turn:85->144, speed:175->248
tricon_zephyr_engine.xml: speed:345->287, accel:195->509, turn:160->213
tricon_zephyr_ii_engine.xml: speed:290->259, accel:160->348, turn:140->170
nexus_dart_160_engine.xml: speed:300->313, accel:165->356, turn:140->192
```
</details>

### The new speed curve
<details><summary>Up to date speed curve</summary>(requires either a light or dark background)

![engines_plot](https://github.com/user-attachments/assets/b28a6038-bda8-47ee-9513-56d18ad814c6)
</details>

## Constraints and technical details

By **max speed**, I mean the speed you have when thrusting.
By **ratio**, I mean the portion of max speed caused by thrusting.
The PHYSICS\_SPEED\_DAMP is used for changing how reactive the ships are, i.e. how their speed changes fast.

The actual accel of the ship is therefore proportional to max\_speed * ratio * PHYSICS\_SPEED\_DAMP.
The actual drift speed of the ship is proportional to max\_speed * (1.0-ratio)
The ratio is used as a measure of agility, therefore turn radius is a function of r (and no other parameter). The turn rate is then deduced from the max speed and turn radius.

Any commit is supposed to have its engine data up to date with parameters values. If not, please report it!

### Constraints - Summary

I considered the following points so far (in addition of the original constraints):
 - avoid the drift speed collapse phenomenon (constraint 1. below) (kept a minimum of +-60% drift)
 - avoid making the Tricon too powerful compared to the other lines (constraint 3. below)
 - keep a reasonable agility ratio between lines (2 from Melendez to Tricon)
 - keep a 2 agility ratio from large to medium and from medium to small.
 - size promotion test (see later post)

I tried to use the full scale of playable behaviors. For example, the Mammoth XL is maximally clunky but still playable, and the zephyr is really agile but yet in the accel and turn limits. Normally this should make a good comparison point for the next steps.

The two points I will check :
 - Do still size 4+ ships feel too clunky ?
   - To me, it is okay. I am waiting for feedback.
 - Are still size 1 ships too agile leading to melee ?
   - Decreased agility magnitude even further. Should be better. -> Seems OK to me.

### Constraints

> However, with the PHYSICS_SPEED_DAMP=3, this makes it feel really awkward when you're not accelerating, since you basically come to a stop. Which is why I lowered it to 1/0.15, which feels quite a bit better, as fighters no longer come to a stop when you don't accel, so you can do better maneuvers with drifting.

This is caused by the ratio getting too high with your formula and this constant setting, therefore the drift speed collapses. We have a first constraint:
 1. Let us keep the ratio below 50% -> meaning max 1.66 for the Tricon ratio.

> The way I approached it was to try to get small ship dogfighting to feel like it was meaningful. So that you can dodge and have the ship feel very responsive. For small ships, something like 600-800 accel feels to be a fairly good spot, which is a lot more than the main values, and also significantly more than this PR.
 2. Around 600-800 accel for small ships. (meaning: interceptors ? -> I assume so)

> However, having Tricon now have a 2:1 ratio, makes it feel sooo much better than the other engines, that, combined with the best top speed, makes the choice of engine a no-brainer. For this reason, it is necessary to tweak the other engines to make them reasonable choices in some situations.
 3. Don't make Tricon a no-brainer choice.

> Additionally, while I was pretty happy with Tricon on size 1-3, it feels weird on size 4+.

 | | speed |	max speed |	accel | fullsp time | fullsp dist | turn | turn radius |
 | --- | --- | --- | --- | --- | --- | --- | --- |
 |T. Zephyr Engine	|335|	401|	440|	0.91|	183|	182|	126|
 |T. Zephyr II Engine	|300|	357|	380|	0.94|	168|	160|	128|
 |T. Cyclone Engine	|260|	305|	300|	1.02|	155|	133|	131|
 |T. Cyclone II Engine	|210|	242.25|	215|	1.13|	136|	105|	132|
 |T. Typhoon Engine	|167|	187.7|	138|	1.36|	128|	76|	142|
 |T. Typhoon II Engine|	127|	138.4|	76|	1.82|	126|	51|	155|

> OK, so I've gone back to the drawing board and tried the three different branches: main, this PR, and my latest changes flying both Lancelot + Admonisher in the gauntlet, and honestly, I feel like I was probably wrong and may be going too overboard.

I don't understand what it means.

> So in main, while accel is very low (like 120 accel for a 330 speed corvette), the advantage of the low accel is that it gives a longer range dimension to combat, which I think is appreciated, and having too high accel (combined with high turn), does make combat a bit too close range and frantic (especially for smaller ships).

I agree with that. That is why Tricon is good for melee and Melendez is nice for bombers (nice drift speed + high engine mass limit to carry all that bombs)

## Guidelines
(Copied text, not mine, even if I agree)

The main thing is figuring out some values. Basically questions like:
 1. Is this accel/speed ratio good for a this class/engine?
 2. Is there enough difference between interceptor / fighter / etc...
 4. Do bulky ships feel bulky? Are NPCs fine with the new values?
 5. Is combat more enjoyable now?
 6. Is traveling more enjoyable now? etc...
 7. Are there actual engine choices or is it just Tricon on bust?