@bobbens 2025-04-01
> The actual accel of the ship is therefore proportional to max_speed * ratio * PHYSICS_SPEED_DAMP. The actual drift speed of the ship is proportional to max_speed * (1.0-ratio) The ratio is used as a measure of agility, therefore turn radius is a function of r (and no other parameter). The turn rate is then deduced from the max speed and turn radius.

Sounds good.

> I can't test, as I had to set the physical constant to 4.5 and no compiled version of Naev has that.

Naev should read PHYSICS_SPEED_DAMP from `dat/constants.lua` now. The value C-side is just the default that gets overwritten (and used if it fails to find the constants file). So you should be able to test with it.

> > OK, so I've gone back to the drawing board and tried the three different branches: main, this PR, and my latest changes flying both Lancelot + Admonisher in the gauntlet, and honestly, I feel like I was probably wrong and may be going too overboard.
> 
> I don't understand what it means.

It means that I probably was chasing too much number tweaking, and lost a bit of a more global reference. Let me elaborate on the points below.

> I considered the following points so far (in addition of the original constraints):
>     * having zephyr accel in [600,800]

After lots of experimenting. I feel like actually going much over a 1 ratio is bad, not because you can dodge better (which is good), but because it makes it so that combat tends to be much closer and ships can chase you down very well, especially with the tight turn radius.

I would suggest lowering the max ratio to not much past 1 for the zephyr, and also lower turn, otherwise enemy fighters become really unfun to fight because they can basically stick to you at near point blank with high accel + turning. Having lower ratio + higher turn radius encourages more a bit more long range combat and keeps things interesting. Potentially makes sense to nerf engine reroute.

So I guess you could consider having combat not end up point blank as a constraint (mainly applicable to small ships).

>     * avoid the drift speed collapse phenomenon (kept a minimum of +-60% drift)

It doesn't feel bad currently.

>     * avoid making the Tricon too powerful compared to the other lines
>     * keep a reasonable agility ratio between lines (2 from Melendez to Tricon)
>     * keep a 2 agility ratio from large to medium and from medium to small.

OK, some impressions of different ships (only testing via Crimson Gauntlet):

1. Lancelot: Generally feels OK without movement upgrades. With 2x engine reroutes, it feels too agile with everything.
   1. + Tricon => flies well, no problem other then issues mentioned about accel + turn above (which are more about breaking combat than how it flies)
   2. + Unicorp => feels fine too, surprisingly easy to fly
   3. + Melendez => also feels quite good, with 2x turn modifier outfits it flies very well, a bit clunky otherwise, but usable,  as should be
2. Pirate Admonisher: feels okish
4. Pirate Kestrel + Krain -> feels quite clunky, not sure if it should feel this clunky. Seems like most stats are similar, but turn has gone from 88 to around 55.
5. Goddard: feels really clunky, I understand having it be clunky to some point, but even with Tricon II + 2x engine reroutes, you can't dodge or really do any maneuvers other than target and tank. I do think if you invest enough into movement it should probably be able to do a bit more with tricons. 
6. Starbridge + Krain -> doesn't feel that bad
7. Pacifier + Tricon -> doesn't feel that bad

Summarizing a bit my impressions:
1. size 1 ships are too agile leading to melee
2. size 2 ships feel a bit too agile, but not that bad, would be a better reference for size 1 ships maybe
3. size 3 ships don't feel that bad, although do have a bit of trouble dodging already
4. size 4+ ships feel potentially too clunky, and in many cases, even with movement upgrades don't feel very good

I would like some other people to test for a second (or third+)  opinion. I also haven't tried normal gameplay nor traveling, only have been focusing on the Gauntlet for now.

Also, as another potential design constraint, I think having it so if you are flying a size X ship, but invest fully in movement upgrades, it may make sense to have it fly like a size X-1 ship. So, say you grab a corvette, remove all weapons, add as many engine reroutes / stabilizers as fits, put the fastest engine, then it should be comparable to a lancelot with no movement upgrades and a similar engine (e.g., both tricon).

> > So in main, while accel is very low (like 120 accel for a 330 speed corvette), the advantage of the low accel is that it gives a longer range dimension to combat, which I think is appreciated, and having too high accel (combined with high turn), does make combat a bit too close range and frantic (especially for smaller ships).
> 
> I agree with that. That is why Tricon is good for melee and Melendez is nice for bombers (nice drift speed + high engine mass limit to carry all that bombs)

I do think it feels a bit like the choice now is mainly Krain vs Tricon. I find Krain the nicest to use (when it fits, which probably should be a more important equipping choice). While the mass limit seems important, the energy requirements feel ignorable at the current time, since worst case you slap on photovoltaic coating + a reactor.